+++ test/test_helper.exs
@@ -1,1 +1,1 @@
-ExUnit.start(exclude: [:requires_lid_176])
+ExUnit.start(exclude: [:requires_lid_176, :requires_bumblebee_model])